requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.4.0
autopep8==2.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4.1
cryptography==37.0.2
decorator==5.1.0
Django==4.0
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
execnet==1.9.0
flake8==6.0.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.2
jsonpickle==2.0.0
MarkupSafe==2.0.1
mccabe==0.7.0
model-bakery==1.7.0
packaging==21.2
pluggy==1.0.0
py==1.11.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.3.0
pyparsing==2.4.7
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-forked==1.4.0
pytest-mock==3.10.0
pytest-xdist==2.5.0
python-decouple==3.5
pytz==2021.3
redis==4.4.2
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.7
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.7
validators==0.18.2

